Optimised gpio bank and read value in int for invalid check

+++ src/GPSU/Components/MCP23017/MCPDevice.hpp
@@ -61,7 +61,7 @@ private:
   void handleWriteEvent(currentEvent *ev);
   void handleSettingChangeEvent(currentEvent *ev);
   void handleBankModeEvent(currentEvent *ev);
-  uint8_t read_mcp_register(const uint8_t reg);
+  int read_mcp_register(const uint8_t reg);
   void write_mcp_register(const uint8_t reg, uint8_t value);
 
   void read_mcp_registers_batch(uint8_t startReg, uint8_t *data, size_t length);
